+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Added
 * Raster implements Clone and Debug
+* Raster::set_rect
 ### Changed
 * Updated to rust 2018 edition